pyproject.toml
[tool.poetry]
name = "county-health-snapshots"
version = "0.1.0"
description = ""
authors = ["Ian <iancharleswatt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"
pyyaml = "^5.1"
requests = "^2.22"
tablib = "^0.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
